Create a minimally viable branch.
- Removes all async features and attempts to simplify design.
